requirements.txt
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlescapy==1.0.1
tzdata==2024.1
